refactor: extract common test logic into abstract base classes
JavaVersionMaven and JavaVersionGradle hold the shared lifecycle
(build, resource, apply + JDK verification, undeploy). Concrete
test classes now only provide project/application names and the
framework-specific log marker.
